Move upload dropdown to separate file
