+++ src/components/editor/EditorComponent.tsx
@@ -4,12 +4,7 @@ import {
   extendedMdRenderer,
 } from './markdown-it-plugged-parser';
 import clsx from 'clsx';
-import { DropDown } from '../library/DropDown';
-import { Button } from '../library/Button';
-import { Icon } from '../library/Icon';
-import { useForm } from 'react-hook-form';
-import { addFileStorageUserIdPost } from '@/api';
-import { serverUrl } from '@/main';
+import { UploadDropdown } from './UploadDropdown';
 
 interface EditorComponentProps extends HTMLAttributes<React.FC> {
   state: string;
@@ -17,10 +12,6 @@ interface EditorComponentProps extends HTMLAttributes<React.FC> {
   active?: boolean;
   extended?: boolean;
   placeholder?: string;
-}
-
-interface UploadFormData {
-  file: FileList;
 }
 
 export const EditorComponent: React.FC<EditorComponentProps> = ({
@@ -34,74 +25,12 @@ export const EditorComponent: React.FC<EditorComponentProps> = ({
     () => (extended ? extendedMdRenderer : simpleRenderer),
     [extended]
   );
-  const { handleSubmit, register } = useForm<UploadFormData>();
-  const [uploadedLink, setUploadedLink] = useState<string | undefined>(
-    undefined
-  );
 
   return (
     <>
       {active ? (
         <div className="w-full h-full">
-          {extended && (
-            <DropDown
-              buttonComponent={
-                <Button title="Upload file for card" variant="bordered">
-                  <Icon icon="upload" />
-                </Button>
-              }
-            >
-              <form
-                className={clsx(
-                  'w-full p-2 vstack',
-                  'bg-1-8 text-black',
-                  'border border-black rounded'
-                )}
-                onSubmit={handleSubmit((data) => {
-                  addFileStorageUserIdPost({
-                    path: {
-                      user_id: 1,
-                    },
-                    body: {
-                      file: data.file[0],
-                    },
-                  })
-                    .then((response) => {
-                      if (response.response.ok)
-                        setUploadedLink(serverUrl + response.data?.url);
-                      else setUploadedLink(response.error?.detail?.toString());
-                    })
-                    .catch((error) => {
-                      console.log(error);
-                    });
-                })}
-              >
-                <div className="around">
-                  <input
-                    type="file"
-                    {...register('file', { required: true })}
-                  />
-                  <Button variant="plate" type="submit">
-                    Upload
-                  </Button>
-                </div>
-                {uploadedLink && (
-                  <div
-                    className="mt-2 around hover:cursor-pointer text-sm"
-                    onClick={() => {
-                      console.log('Copying ' + uploadedLink);
-                      navigator.clipboard.writeText(uploadedLink);
-                    }}
-                  >
-                    <span className="mr-1">Link to file (click to copy):</span>
-                    <span>
-                      <u>{uploadedLink}</u>
-                    </span>
-                  </div>
-                )}
-              </form>
-            </DropDown>
-          )}
+          {extended && <UploadDropdown />}
           <textarea
             className={clsx(
               'bg-1-2 focus:bg-1-3',
